Allow http-rb gem 5.x
the `http` gem released 5.0.0 on May 13 2021. https://rubygems.org/gems/http/versions/5.0.0

All tests in this repo pass with http 5.0.

I opted to leave both http 4.x and 5.x supported in gemfile, instead of forcing upgrade to 5.x. this seems kinder to users, with less chance of causing dependency incompatibilities in an app's entire dependency tree, and without arguably requiring a major version bump to faraday-http if it dropped support for http 4.